Test Windows gem against encrypted MSSQL connections
The tiny_tds version binary we build for Windows is compiled against OpenSSL. Therefore, if we make any changes there (e.g. version or compilation options), it would be good to check that a connection against a MSSQL server which requires encryption still works.

For Linux, I do think this is necessary as there everybody compiles their own version of freetds with openssl.